refactor: use another approach for fixing Safari bug
