+++ src/shared.ts
@@ -23,9 +23,8 @@ export const _DOMException =
   typeof DOMException === 'object' ? Error : DOMException;
 
 export const isSafari =
-  // https://github.com/hgoebl/mobile-detect.js/blob/master/mobile-detect.js#L191
-  /Version((?!\bEdgiOS\b).)*Mobile.*Safari|Safari.*Mobile|MobileSafari/.test(
-    navigator.userAgent,
-  );
+  /Safari\//.test(navigator.userAgent) &&
+  !/Chrom(e|ium)\//.test(navigator.userAgent);
 
-export const safariBrokenRules = ['content'];
+export const defineProperty = Object.defineProperty;
+export const forEach = Array.prototype.forEach;
