+++ src/utils.ts
@@ -1,7 +1,5 @@
-import {closedShadowRootRegistry, safariBrokenRules} from './shared';
-
-export const defineProperty = Object.defineProperty;
-export const forEach = Array.prototype.forEach;
+import {removeSafariPlaceholder} from './safari';
+import {closedShadowRootRegistry, forEach, isSafari} from './shared';
 
 const importPattern = /@import.+?;?$/gm;
 
@@ -25,7 +23,10 @@ export function clearRules(sheet: CSSStyleSheet): void {
 
 export function insertAllRules(from: CSSStyleSheet, to: CSSStyleSheet): void {
   forEach.call(from.cssRules, (rule, i) => {
-    to.insertRule(rule.cssText, i);
+    to.insertRule(
+      isSafari ? removeSafariPlaceholder(rule.cssText) : rule.cssText,
+      i,
+    );
   });
 }
 
@@ -58,60 +59,4 @@ export function removeNode(node: Node): void {
 
 export function getShadowRoot(element: Element): ShadowRoot | undefined {
   return element.shadowRoot || closedShadowRootRegistry.get(element);
-}
-
-export function escapeRegexString(str: string): string {
-  return str.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&').replace(/-/g, '\\x2d');
-}
-
-// WeakSet emulation for IE11.
-const processedRulesRegistry = new WeakMap<CSSRule, boolean>();
-/**
- * This function goes through all the broken Safari rules that are specified in
- * `shared.ts`. If the user had an intention to add quotes to the value of
- * broken rule, this function detects it and forcefully add quotes to the value
- * of the rule declaration.
- *
- * @param sheet The basic style sheet
- * @param originalStyle The style string user provided
- */
-export function fixSafariBrokenRules(
-  sheet: CSSStyleSheet,
-  originalStyle: string,
-) {
-  safariBrokenRules.forEach((ruleName) => {
-    // Detect if the user has declared a rule with the specified name, then run
-    // the fixing algorithm.
-    if (originalStyle.indexOf(`${ruleName}:`) > -1) {
-      forEach.call(sheet.cssRules, (rule) => {
-        // Skip if the rule is already fixed
-        if (processedRulesRegistry.has(rule)) {
-          return;
-        }
-
-        // Safari stores the value without quotes, e.g.,
-        // { "content": "some string" } instead of
-        // { "content": "\"some string\"" like other browsers do.
-        const ruleValue = rule.style[ruleName];
-        if (ruleValue !== '') {
-          // Detecting if the user intended to add quotes here.
-          const pattern = new RegExp(
-            `${escapeRegexString(
-              rule.selectorText,
-            )}(?:\\s|.)*${escapeRegexString(
-              ruleName,
-            )}\\s*(['"]?)${escapeRegexString(ruleValue)}["']?`,
-          );
-
-          const matches = pattern.exec(originalStyle)!;
-
-          if (matches[1] !== '') {
-            rule.style[ruleName] = `${matches[1]}${ruleValue}${matches[1]}`;
-          }
-
-          processedRulesRegistry.set(rule, true);
-        }
-      });
-    }
-  });
 }
